rm anchor from labs
