+++ packages/gamut-labs/src/experimental/index.ts
@@ -1,4 +1,3 @@
-export * from './Anchor';
 export * from './Card';
 export * from './Popover';
 export * from './Text';
