+++ packages/styleguide/stories/Core/Typography/Anchor.stories.mdx
@@ -1,4 +1,4 @@
-import { Anchor } from '@codecademy/gamut-labs/src';
+import { Anchor } from '@codecademy/gamut';
 import { Text } from '@codecademy/gamut/src';
 import { Meta, Story, Canvas, Props } from '@storybook/addon-docs/dist/blocks';
 import { VisualTheme } from '@codecademy/gamut/src';
@@ -7,8 +7,8 @@ import { VisualTheme } from '@codecademy/gamut/src';
   title="Typography/Anchor"
   component={Anchor}
   parameters={{
-    status: 'volatile',
-    source: 'gamut-labs',
+    status: 'stable',
+    source: 'gamut',
   }}
 />
 
@@ -63,11 +63,11 @@ We provide 2 style variants for Anchor.
 </Canvas>
 
 <Canvas>
-  <Story name="Not Inline" args={{ variant: 'notInline' }}>
+  <Story name="Interface" args={{ variant: 'interface' }}>
     {(args) => {
       return (
         <Text>
-          This non-inline style should only be used when link is not part of the
+          This interface style should only be used when link is not part of the
           body text and it doesn't make sense to use Text Button.
           <br />
           <br />
@@ -101,4 +101,4 @@ For general links in body text, we use a visual underline to make it clear that 
   <Anchor href="/">This is a link</Anchor>. This is not.
 </Text>
 
-Only use `notInline` variant when your link text is clearly separated from body text in a way that implies a link. Otherwise, use the default `inline` Anchor style; remember [color contrast](https://www.w3.org/WAI/WCAG21/Understanding/contrast-minimum.html) and [non-color visual distinction](https://www.w3.org/TR/WCAG20-TECHS/F73.html): users who are color-blind and/or low vision should still be able to see the link and text distinction.
+Only use `interface` variant when your link text is clearly separated from body text in a way that implies a link. Otherwise, use the default `inline` Anchor style; remember [color contrast](https://www.w3.org/WAI/WCAG21/Understanding/contrast-minimum.html) and [non-color visual distinction](https://www.w3.org/TR/WCAG20-TECHS/F73.html): users who are color-blind and/or low vision should still be able to see the link and text distinction.
